Extract legacy NodeDatabase migration
